# .stage0_template/test_expected/docs/openapi.yaml
openapi: 3.0.3
info:
  title: "partner API"
  description: |-
    This is a super simple API for managing partners
  contact:
    email: "devs@agile-learning-institute"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11

paths:


  /api/partner/:
    get:
      summary: Get a list of partner's names and ids
      operationId: get_partner
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/partner"
        "500":
          description: An error has occurred 
    post:
      summary: "Add a new partner"
      operationId: "addpartner"
      requestBody:
        description: "partner details to be added"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/partner"
              required: true
      responses:
        "200":
          description: "partner added successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/partner"
        "500":
          description: An error has occurred 

  /api/partner/partnerId}/:
    get:
      summary: "Get an existing partner by ID"
      operationId: "getpartnerById"
      parameters:
        - name: "partnerId"
          in: path
          description: The ID of the partner to retrieve
          required: true
          schema:
            type: string
            format: GUID
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/partner"
        "500":
          description: An error has occurred 
    patch:
      summary: Update a partner by ID
      operationId: updatepartner
      parameters:
        - name: "partnerId"
          in: path
          description: "The ID of the partner to update"
          required: true
          schema:
            type: string
            format: GUID
      requestBody:
        description: "updated partner details"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/partner"
        required: true
      responses:
        "200":
          description: "partner updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/partner"
        "500":
          description: An error has occurred 



  /api/partner/:
    get:
      summary: Get a list of partner 
      operationId: getpartner
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: The string to search for
          required: true
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/partner"
        "500":
          description: An error has occurred 


  /api/config/:
    get:
      summary: Get API Configuration Information
      operationId: getConfig
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/config"

  /api/health/:
    get:
      summary: Prometheus Healthcheck endpoint
      operationId: getHealth
      responses:
        "200":
          description: Successful operation

components:
  schemas:

    partner:
      $ref:./dataDefinitions/dd.partner.yaml



    person:
      $ref:./dataDefinitions/dd.person.yaml


    config:
      type: object
      properties:
        configItems:
          type: array
          items:
            type: object
            properties:
              name:
                description: Conf Item Name (Env Var Name, File Name)
                type: string
              value:
                description: The value for that config item
                type: string
              from:
                description: Where the value was found
                type: string
                enum:
                  - default
                  - environment
                  - file
        versions:
          type: array
          items:
            type: object
            properties:
              collectionName:
                description: Mongodb Collection name
                type: string
              currentVersion:
                description: Schema Version for the collection
                type: string
        enumerators:
          type: object
        mentors:
          description: List of all members with role Mentor
          type: array
          items:
            type: object